python-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

sentence-transformers==3.3.1
mistralai==1.5.1
groq==0.15.0
numpy==2.2.1
scikit-learn==1.6.1
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.5
